requirements.txt
# Core dependencies
anthropic>=0.34.0
claude-agent-sdk>=0.1.18

# Async & utilities
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
